# loyalty-platform-impl/openapi.yaml
openapi: "3.0.3"
info:
  title: Loyalty Banking Platform — In-Scope API (Capstone)
  description: |
    Public contract for the four I-11 use cases (UC-LB-01 through UC-LB-04).
    Scope: I-11 only. Names are Lab 1 I-4 container names.
    I-3 external systems (Core Banking System, Partner Systems, CRM & Notification Gateway,
    Enterprise Data Warehouse) are simulated at the integration boundary; no production credentials
    or real host names are used.

    **Constraint reference:**
    - CON.1 — No duplicate point posting for the same source transaction
    - CON.2 — No direct database writes from channels or other services
    - CON.3 — Fulfillment failure must compensate by restoring points (FIFO earn date and expiry preserved)
    - CON.4 — Analytics reporting data must not exceed 10 minutes staleness

    **Spec source:** lab3-spec.md §4 contract register (CT-04, CT-11, CT-12, CT-26)
    **RACI:** R Dev (Lê Huy Du) · A SA (Vũ Trường Quang) · C Test (Lê Huy Du)
  version: "1.0.0"
  contact:
    name: Team 2 — Vũ Trường Quang, Khuất Duy Bách, Đặng Duy Hoàng, Lê Huy Du
servers:
  - url: http://localhost:8081
    description: Earning Engine Service (I-4, Domain Services Zone)
  - url: http://localhost:8082
    description: Tiering System Service (I-4, Domain Services Zone)
  - url: http://localhost:8083
    description: Redemption Engine Service (I-4, Domain Services Zone)
  - url: http://localhost:8085
...
paths:
...
  /api/v1/redemptions/orders:
    post:
      operationId: createRedemptionOrder
      tags:
        - UC-LB-02 Redemption Engine Service
      summary: UC-LB-02 happy path — place a redemption order with FIFO debit
      description: |
        Happy path: Member submits order via API Gateway → Redemption Engine Service.
        Modules M2–M7 (lab3-spec.md §2): lock balance (M4), check tier + minimum (M3), debit FIFO (M5),
        move RedemptionOrder PENDING → IN_PROGRESS (M6), dispatch fulfillment (M7).

        Named alts:
        - **ALT-01 / ERR_RED_INSUFFICIENT_BALANCE** — available balance < reward cost → 422 (G6-A02, T-06)
        - **ALT-02 / ERR_RED_TIER_ELIGIBILITY_FAILED** — MemberTier below minimum → 422 (G6-A02, T-07)
        - **ALT-03 / ERR_RED_MIN_POINTS** — request below 100-point minimum → 422 (T-08)
        - **ALT-06 / ERR_RED_CONCURRENT_REQUEST** — balance lock busy → 409 (T-11)
        - **ALT-05 / CON.3** — partner fulfillment failure → REVERSED async (EXC-05, G6-A03, T-10):
          order moves IN_PROGRESS → FAILED → REVERSED; points restored with original FIFO earn date and expiry.
          Notified via PATCH /api/v1/redemptions/orders/{orderId}/fail.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedemptionOrderRequest'
            example:
              memberId: "member-001"
              programId: "DEFAULT_PROG"
              rewardItemId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
              quantity: 1
      responses:
        "201":
          description: |
             Created. RedemptionOrder is in IN_PROGRESS after authoritative CT-12 eligibility
             and CT-13 FIFO debit succeed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionOrderResponse'
        "409":
          description: |
            Conflict — ALT-06 balance lock is busy (ERR_RED_CONCURRENT_REQUEST, T-11).
            Concurrent request holds the lock; retry shortly.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://loyalty.internal/errors/concurrent-request"
                title: "Balance Lock Busy"
                status: 409
                detail: "ERR_RED_CONCURRENT_REQUEST: Balance lock busy, retry shortly"
                instance: "/api/v1/redemptions/orders"
        "422":
          description: |
            Unprocessable — one of the validation alts fired:
            - ALT-01: ERR_RED_INSUFFICIENT_BALANCE — balance < reward cost (T-06)
            - ALT-02: ERR_RED_TIER_ELIGIBILITY_FAILED — MemberTier below minimum (T-07)
             - ALT-03: ERR_RED_MIN_POINTS — below 100-point minimum (T-08)
             No points are debited in any of these cases; no order is persisted before debit succeeds.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "400":
          description: Bad request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/redemptions/orders/{orderId}/fulfill:
    patch:
      operationId: fulfillRedemptionOrder
      tags:
        - UC-LB-02 Redemption Engine Service
      summary: UC-LB-02 — partner confirms delivery; order moves IN_PROGRESS → FULFILLED (G6-T03)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order FULFILLED. Debit is final. Happy path terminal state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionOrderResponse'
        "404":
          description: Order not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/redemptions/orders/{orderId}/fail:
    patch:
      operationId: failAndReverseRedemptionOrder
      tags:
        - UC-LB-02 Redemption Engine Service
      summary: |
        UC-LB-02 alt — partner fulfillment fails; order moves IN_PROGRESS → FAILED → REVERSED (CON.3, ALT-05, EXC-05, G6-T04/T05, G6-A03)
      description: |
        Named alt ALT-05 / EXC-05 / CON.3:
        Partner Systems return a fulfillment failure after points were debited.
        RedemptionOrder moves IN_PROGRESS → FAILED then FAILED → REVERSED.
        The debited point batches are restored to their original FIFO earn date and expiry.
        Member is notified via CRM & Notification Gateway (mocked).
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "OUT_OF_STOCK"
      responses:
        "200":
          description: |
            Order moved through FAILED → REVERSED. Points restored with original earn date and expiry (CON.3).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionOrderResponse'
        "404":
          description: Order not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/redemptions/orders/{orderId}/cancel:
    patch:
      operationId: cancelRedemptionOrder
      tags:
        - UC-LB-02 Redemption Engine Service
      summary: UC-LB-02 alt — Member cancels while PENDING (ALT-04, G6-T02, T-09)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId]
              properties:
                memberId:
                  type: string
      responses:
        "200":
          description: Order CANCELLED. No points debited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionOrderResponse'
        "409":
          description: |
            Cancellation refused — order is not in PENDING state (T-09 variant).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/catalog/items:
    post:
      operationId: addCatalogItem
      tags:
        - UC-LB-02 Redemption Engine Service
      summary: UC-LB-02 — add reward catalog item (M1 Catalog Query setup, CT-11 QueryCatalog)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardItemRequest'
      responses:
        "201":
          description: Catalog item created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardItemResponse'

  # ─────────────────────────────────────────────────────────────────
  # UC-LB-03 — Apply tier upgrade (CT-12 GetMemberTier)
  # Server: http://localhost:8082
  # ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RedemptionOrderRequest:
      type: object
      required: [memberId, rewardItemId]
      properties:
        memberId:
          type: string
          example: "member-001"
        programId:
          type: string
          default: "DEFAULT_PROG"
          example: "DEFAULT_PROG"
        rewardItemId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          default: 1
          example: 1

    # ── CT-13 FIFO Debit & Restoration Schemas ───────────
...
    RedemptionOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        memberId:
          type: string
        programId:
          type: string
        rewardItemId:
          type: string
          format: uuid
        totalPointsDebited:
          type: integer
        memberTierAtOrder:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, FULFILLED, FAILED, REVERSED, CANCELLED]
          description: |
            I-6 RedemptionOrder states. Terminal states: CANCELLED, FULFILLED, REVERSED.
            State machine per lab10 UC-LB-02 and lab3-spec.md §2 M6 Order State Keeper.
        failureReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    RewardItemRequest:
      type: object
      required: [programId, name, category, pointsCost, fulfillmentType, minTierRequired]
      properties:
        programId:
          type: string
        name:
          type: string
        category:
          type: string
        pointsCost:
          type: integer
          minimum: 100
          description: Must be >= 100 (minimum points constraint FR-03-013).
        currencyValue:
          type: number
          format: double
        fulfillmentType:
          type: string
          enum: [DIGITAL, PHYSICAL]
        minTierRequired:
          type: string
          enum: [SILVER, GOLD, PLATINUM]
    RewardItemResponse:
      type: object
      properties:
        itemId:
          type: string
          format: uuid
        programId:
          type: string
        name:
          type: string
        pointsCost:
          type: integer
        minTierRequired:
          type: string
        status:
          type: string

    # ── UC-LB-03 ─────────────────────────────────────────
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem detail.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string